+++ src/main/java/edu/ucsb/cs156/happiercows/HappierCowsApplication.java
@@ -1,13 +1,13 @@
-package edu.ucsb.cs156.kitchensink;
+package edu.ucsb.cs156.happiercows;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class KitchenSinkApplication {
+public class HappierCowsApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(KitchenSinkApplication.class, args);
+    SpringApplication.run(HappierCowsApplication.class, args);
   }
 
 }
