+++ src/main/java/edu/ucsb/cs156/happiercows/controllers/UserInfoController.java
@@ -1,8 +1,5 @@
-package edu.ucsb.cs156.kitchensink.controllers;
+package edu.ucsb.cs156.happiercows.controllers;
 
-import edu.ucsb.cs156.kitchensink.entities.User;
-import edu.ucsb.cs156.kitchensink.models.CurrentUser;
-import edu.ucsb.cs156.kitchensink.services.CurrentUserService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 
@@ -10,6 +7,10 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import edu.ucsb.cs156.happiercows.entities.User;
+import edu.ucsb.cs156.happiercows.models.CurrentUser;
+import edu.ucsb.cs156.happiercows.services.CurrentUserService;
 
 @Api(description="Current User Information")
 @RequestMapping("/api/currentUser")
