+++ src/main/java/edu/ucsb/cs156/happiercows/models/CurrentUser.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs156.kitchensink.models;
+package edu.ucsb.cs156.happiercows.models;
 
 import lombok.Data;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,8 @@ import lombok.AccessLevel;
 
 
 import org.springframework.security.core.GrantedAuthority;
-import edu.ucsb.cs156.kitchensink.entities.User;
+
+import edu.ucsb.cs156.happiercows.entities.User;
 
 import java.util.Collection;
