+++ src/test/java/edu/ucsb/cs156/happiercows/ControllerTestCase.java
@@ -1,11 +1,12 @@
-package edu.ucsb.cs156.kitchensink;
+package edu.ucsb.cs156.happiercows;
 
-import edu.ucsb.cs156.kitchensink.entities.User;
-import edu.ucsb.cs156.kitchensink.services.CurrentUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
+
+import edu.ucsb.cs156.happiercows.entities.User;
+import edu.ucsb.cs156.happiercows.services.CurrentUserService;
 
 import static org.mockito.Mockito.*;
