+++ src/test/java/edu/ucsb/cs156/happiercows/HappierCowsApplicationTests.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs156.kitchensink;
+package edu.ucsb.cs156.happiercows;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,7 +6,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 @ActiveProfiles("test")
 @SpringBootTest
-class KitchenSinkApplicationTests {
+class HappierCowsApplicationTests {
 
   @Test
   void contextLoads() {
