+++ src/test/java/edu/ucsb/cs156/happiercows/controllers/UserInfoControllerTests.java
@@ -1,7 +1,8 @@
-package edu.ucsb.cs156.kitchensink.controllers;
+package edu.ucsb.cs156.happiercows.controllers;
 
-import edu.ucsb.cs156.kitchensink.ControllerTestCase;
-import edu.ucsb.cs156.kitchensink.repositories.UserRepository;
+import edu.ucsb.cs156.happiercows.ControllerTestCase;
+import edu.ucsb.cs156.happiercows.controllers.UserInfoController;
+import edu.ucsb.cs156.happiercows.repositories.UserRepository;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
